requirements.txt
appdirs==1.4.4
asgiref==3.5.2
Babel==2.10.3
beautifulsoup4==4.11.1
bootstrap4==0.1.0
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
confusable-homoglyphs==3.2.0
contourpy==1.0.5
cryptography==38.0.1
cssselect==1.2.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==3.2.9
django-allauth==0.41.0
django-appconf==1.0.5
django-auto-logout==0.5.0
django-bootstrap4==22.2
django-bower==5.2.0
django-crispy-forms==1.14.0
django-embed-video==1.4.7
django-heroku==0.3.1
django-import-export==3.0.1
django-on-heroku==1.1.2
django-prices==2.3.0
django-registration==3.3
django-registration-redux==2.4
django-reset-migrations==0.4.0
django-rest-auth==0.9.5
django-select2==6.1.0
django-tables2==2.4.1
django-tables2-column-shifter==2.0.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
enmerkar==0.7.1
et-xmlfile==1.0.1
fake-useragent==0.1.11
fonttools==4.38.0
gunicorn==20.1.0
idna==2.8
importlib-metadata==5.0.0
jdcal==1.4.1
jieba==0.38
kiwisolver==1.4.4
lml==0.0.9
lxml==4.9.1
Markdown==2.6.8
MarkupPy==1.14
matplotlib==3.6.1
numpy==1.23.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
parse==1.19.0
Pillow==9.2.0
prices==1.1.1
psycopg2==2.8.5
psycopg2-binary==2.9.5
pycparser==2.20
pyee==8.2.2
pyexcel-io==0.5.20
Pygments==2.2.0
PyJWT==1.7.1
pylibjpeg-libjpeg==1.3.2
pyparsing==2.4.7
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.3.1
reportlab==3.6.12
requests==2.22.0
requests-html==0.10.0
requests-oauthlib==1.3.0
shutilwhich==1.1.0
six==1.12.0
snowballstemmer==2.0.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.3.2.post1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
tablib==3.2.1
texttable==1.6.2
tqdm==4.64.1
tzdata==2022.5
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.11
w3lib==2.0.1
waitress==2.1.2
webencodings==0.5.1
WebOb==1.8.6
websockets==10.4
whitenoise==5.0.1
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==3.10.0
